requirements.txt
torch
torchvision
transformers
datasets
accelerate
scikit-learn
Pillow
requests
tqdm
